Popup confirmation and Draggable tasks

+++ src/renderer/app.js
@@ -1,15 +1,33 @@
 import { createTask, deleteTask, createList } from '../logic/taskManager.js';
 
 let lists = [];
+let currentList;
 
 window.storage.getTasks().then(loadedLists => {
     lists = loadedLists;
     loadLists();
     if (lists.length === 0) addList();
     loadTaskElements(lists[0]);
+    currentList = lists[0];
 })
 
 window.onload = () => {
+
+    let tasks = document.querySelector('#todo-list');
+
+    Sortable.create(tasks, {
+        animation: 150,
+
+        onEnd({ oldIndex, newIndex }) {
+
+            const moved = currentList.tasks.splice(oldIndex, 1)[0];
+
+            currentList.tasks.splice(newIndex, 0, moved);
+
+            window.storage.saveTasks(lists);
+        }
+    });
+
     let input = document.querySelector('#item-input');
     input.addEventListener('keydown', (e) => {
         if (e.key === 'Enter' && input.value.trim() !== '') {
@@ -37,6 +55,7 @@ window.onload = () => {
 
     let listSelect = document.querySelector('#list-select');
     listSelect.addEventListener('change', (e) => {
+        currentList = lists.find(list => list.id === e.target.value);
         swapLists(listSelect);
     });
 
@@ -45,15 +64,22 @@ window.onload = () => {
         listSelect.value = lists[lists.length - 1].id;
         swapLists(listSelect);
     });
+
     let deleteButtons = document.querySelectorAll('.delete-list');
     for (let i = 0; i < deleteButtons.length; i++) {
-        deleteButtons[i].addEventListener('click', () => {
+        deleteButtons[i].addEventListener('click', async () => {
+            let result = await AskConfirmation(`Are you sure you want to delete ${currentList.title}?`);
+            if (result === 1) return;
             deleteList();
             listSelect.value = lists[0].id;
             swapLists(listSelect);
         });
     }
+}
 
+let AskConfirmation = async (text) => {
+    const result = await window.app.confirmDialog(text);
+    return result;
 }
 
 let checkNoTasks = () => {
@@ -85,7 +111,7 @@ let deleteList = () => {
     lists = lists.filter(list => list.id !== document.querySelector('#list-select').value);
     saveTasks();
     document.querySelector('#list-select').removeChild(document.querySelector('#list-select').selectedOptions[0]);
-    if(lists.length === 0) addList();
+    if (lists.length === 0) addList();
 }
 
 let addList = () => {
@@ -97,6 +123,7 @@ let addList = () => {
     listElement.textContent = newList.title;
     document.querySelector('#list-select').appendChild(listElement);
     showNoTasks();
+    currentList = newList;
 }
 
 let swapLists = (e) => {
@@ -165,7 +192,7 @@ function showTitleInput(element) {
         element.textContent = input.value;
         input.hidden = true;
         element.hidden = false;
-        let currentList = lists.find(list => list.id === document.querySelector('#list-select').value);
+        currentList = lists.find(list => list.id === document.querySelector('#list-select').value);
         currentList.title = input.value;
 
         document.querySelectorAll('option').forEach(option => {
@@ -254,3 +281,4 @@ function saveTasks() {
         checkNoTasks();
     });
 }
+
